CameraControlScreen switch state bug fixed

+++ Sources/Platform/Windows/Native.cpp
@@ -44,24 +44,24 @@ LRESULT CALLBACK WinProc(HWND wnd, UINT msg, WPARAM wParam, LPARAM lParam){
 	switch(msg){
 	case WM_LBUTTONDOWN:{
 		SetCapture(wnd);
-		float targetX = (float)LOWORD(lParam);
-		float targetY = (float)HIWORD(lParam);
+		float targetX = (short)LOWORD(lParam);
+		float targetY = (short)HIWORD(lParam);
 		mouseDown = true;
 		TRIGGER_EVENT(input->TargetActionDown, targetX, targetY);
 		break;
 	}
 	case WM_MOUSEMOVE:{
 		if(mouseDown){
-			float targetX = (float)LOWORD(lParam);
-			float targetY = (float)HIWORD(lParam);
+			float targetX = (short)LOWORD(lParam);
+			float targetY = (short)HIWORD(lParam);
 			TRIGGER_EVENT(input->TargetActionMove,targetX, targetY);
 		}
 		break;
 	}
 	case WM_LBUTTONUP:{
 		ReleaseCapture();
-		float targetX = (float)LOWORD(lParam);
-		float targetY = (float)HIWORD(lParam);
+		float targetX = (short)LOWORD(lParam);
+		float targetY = (short)HIWORD(lParam);
 		mouseDown = false;
 		TRIGGER_EVENT(input->TargetActionUp, targetX, targetY);
 		break;
